Deprecate fake components, long live mock components
